requirements.txt
Flask
pdfplumber
pytesseract
Pillow
opencv-python
numpy<2.0.0
opencv-python-headless<4.10
scikit-learn
flask_sqlalchemy
flask_bcrypt
matplotlib
easyocr
google-genai
reportlab
python-dotenv
sentence-transformers>=2.6.0
transformers<4.40.0
torch
tf-keras
cryptography
